requirements.txt
aiogram==3.7.0
fastapi==0.111.0
uvicorn==0.29.0
anthropic>=0.40.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.0
greenlet>=3.0.0
